Fix circular service dependency injection.
